moved const to const
